llm-plotly/requirements.txt
langchain
langchain_groq
dash
dash_ag_grid
pandas
python-dotenv
